refactor(test): simplify maintenance test support
