+++ apps/agent-worker/package.json
@@ -37,6 +37,7 @@
 	"devDependencies": {
 		"@biomejs/biome": "^2.3.11",
 		"@electric-sql/pglite": "^0.5.3",
+		"@mymemo/test-support": "workspace:*",
 		"@types/bun": "^1.3.11",
 		"@types/pg": "^8.20.0"
 	}
